// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/ratetagrolebranchconfiguration.ts
export class RateTagRoleBranchConfigurationModel {
    rateTagRoleBranchConfigurationId: string;
    branchId: string;
    roleId: string;
    startDate: Date;
    endDate: Date;
    timeStamp: any;
    isArchived: boolean;
    totalCount: number;
    branchName: string;
    roleName: string;
    priority: number;
}