// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/ratetagrolebranchconfigurationinputmodel.ts
export class RateTagRoleBranchConfigurationInputModel {
    rateTagRoleBranchConfigurationId: string;
    branchId: string;
    roleId: string;
    employeeId: string;
    startDate: Date;
    endDate: Date;
    isDisabled: boolean;
    isInHerited: boolean;
    isOverRided: boolean;
    priority: number;
}